Clean up actionWidget
- Reduce number of exports
- Adding private and removing unnused params
- Split out `actionList.ts` to own file
